Fixed messup
I realized I changed the docs for impulseplot, thinking that it was the docs for stepplot, in my last commit. This is now fixed (in accordance with docs for step, not yet in accordance with impulse. Waiting to hear @olof3's thoughs on this phrasing before changing docs for impulse)